% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getQuantLibCapabilities}
\alias{getQuantLibCapabilities}
\title{Return configuration options of the QuantLib library}
\usage{
getQuantLibCapabilities()
}
\value{
A named vector of logical variables
}
\description{
This function returns a named vector of boolean variables describing several
configuration options determined at compilation time of the QuantLib library.
}
\details{
Not all of these features are used (yet) by RQuantLib.
}
\examples{
  getQuantLibCapabilities()
}
\references{
\url{http://quantlib.org} for details on \code{QuantLib}.
}
\author{
Dirk Eddelbuettel
}
