% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tsQuotes}
\alias{tsQuotes}
\title{Vol Cube Example Data  
 
Short time series examples}
\format{A series of tenors and rates approppriate for calling DiscountCurve}
\source{
TBA
}
\description{
Vol Cube Example Data  
 
Short time series examples
}
