% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vcube}
\alias{vcube}
\title{Vol Cube Example Data}
\format{two data frames: \code{vcube}, a data frame with four columns: \code{Expiry}, \code{Tenor},
\code{LogNormalVol},  and \code{Spread}}
\source{
TBA
}
\usage{
data(vcube)
}
\description{
Data for valuing swaption examples including rates and a lognormal vol cube
}
