% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\docType{data}
\name{calendars}
\alias{calendars}
\title{The \code{calendars} vector contains all calendar identifiers.}
\description{
The \code{calendars} vector contains all calendar identifiers.
}
\examples{
head(calendars, 10)
}
\keyword{data}
