% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getQuantLibVersion}
\alias{getQuantLibVersion}
\title{Return the QuantLib version number}
\usage{
getQuantLibVersion()
}
\value{
A character variable
}
\description{
This function returns the QuantLib version string as encoded in the header
file \code{config.hpp} and determined at compilation time of the QuantLib library.
}
\examples{
  getQuantLibVersion()
}
\references{
\url{http://quantlib.org} for details on \code{QuantLib}.
}
\author{
Dirk Eddelbuettel
}
